+++ cardano-cli/src/Cardano/CLI/Byron/Delegation.hs
@@ -5,10 +5,8 @@
 {-# LANGUAGE ScopedTypeVariables        #-}
 {-# OPTIONS_GHC -Wno-all-missed-specialisations #-}
 
-module Cardano.CLI.Delegation
-  ( CertificateFile(..)
-  , NewCertificateFile(..)
-  , issueByronGenesisDelegation
+module Cardano.CLI.Byron.Delegation
+  ( issueByronGenesisDelegation
   , checkByronGenesisDelegation
   )
 where
@@ -29,9 +27,6 @@ import           Cardano.Crypto (ProtocolMagicId)
 import           Cardano.CLI.Ops
 
 
-newtype NewCertificateFile =
-  NewCertificateFile { nFp :: FilePath }
-  deriving (Eq, Ord, Show, IsString)
 
 
 -- TODO:  we need to support password-protected secrets.
