+++ cardano-cli/src/Cardano/CLI/Byron/Genesis.hs
@@ -7,7 +7,7 @@
 #define UNIX
 #endif
 
-module Cardano.CLI.Genesis
+module Cardano.CLI.Byron.Genesis
   ( NewDirectory(..)
   , GenesisParameters(..)
   , dumpGenesis
@@ -45,7 +45,7 @@ import           Cardano.Chain.Genesis (GeneratedSecrets(..))
 import           Cardano.Crypto (SigningKey (..))
 import qualified Cardano.Crypto as Crypto
 
-import           Cardano.CLI.Key
+import           Cardano.CLI.Byron.Key
 import           Cardano.CLI.Ops
