+++ cardano-cli/src/Cardano/CLI/Byron/UpdateProposal.hs
@@ -37,11 +37,12 @@ import qualified Ouroboros.Consensus.Mempool as Mempool
 import           Ouroboros.Consensus.Node.ProtocolInfo (pInfoConfig)
 import           Ouroboros.Network.NodeToClient (IOManager)
 
-import           Cardano.CLI.Key (readEraSigningKey)
+import           Cardano.CLI.Byron.Key (readEraSigningKey)
 import           Cardano.CLI.Ops
                    (CardanoEra(..), CliError(..), ensureNewFileLBS,
                     readGenesis, withRealPBFT)
 import           Cardano.Common.LocalSocket
+
 
 runProposalCreation
   :: ConfigYamlFilePath
