+++ cardano-cli/src/Cardano/CLI/Shelley/Run/Pool.hs
@@ -26,7 +26,6 @@ import qualified Shelley.Spec.Ledger.Slot as Shelley
 
 import           Cardano.Config.Shelley.VRF
 
-import           Cardano.CLI.Key (VerificationKeyFile(..))
 import           Cardano.CLI.Ops (CliError (..))
 import           Cardano.CLI.Shelley.Commands
